+++ modules/main/src/main/java/com/nisovin/shopkeepers/compat/Compat.java
@@ -47,6 +47,7 @@ public final class Compat {
 	// Minecraft version instead.
 	static {
 		// Registered in the order from latest to oldest.
+		register(new CompatVersion("26_1_R1", "26.1", "e8ece90188c951d866bd2fffc52c803e"));
 		register(new CompatVersion("1_21_R9_paper", "1.21.11", "1.21.11"));
 		register(new CompatVersion("1_21_R9", "1.21.11", "e3cd927e07e6ff434793a0474c51b2b9"));
 		// 1.21.9: Not supported. Superseded by 1.21.10.